# Tracing: optimize and reduce overhead