--- a/ocaml/libs/tracing/tracing.mli
+++ b/ocaml/libs/tracing/tracing.mli
@@ -120,7 +120,9 @@ module Span : sig
   val get_attributes : t -> (string * string) list
 end
 
-module SpanMap : module type of Map.Make (Trace_id)
+module TraceMap : module type of Map.Make (Trace_id)
+
+module SpanMap : module type of Map.Make (Span_id)
 
 module Spans : sig
   val set_max_spans : int -> unit
@@ -131,7 +133,7 @@ module Spans : sig
 
   val since : unit -> Span.t list * int
 
-  val dump : unit -> Span.t list SpanMap.t * (Span.t list * int)
+  val dump : unit -> Span.t SpanMap.t TraceMap.t * (Span.t list * int)
 end
 
 module Tracer : sig